+++ src/Http/Actions/ForgotPasswordAction.php
@@ -2,14 +2,14 @@
 
 declare(strict_types=1);
 
-namespace Technobase\AuthKit\Http\Callbacks;
+namespace Technobase\AuthKit\Http\Actions;
 
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Support\Facades\Password;
 use Technobase\AuthKit\Http\Requests\ForgotPasswordRequest;
 
-final class ForgotPasswordCallback
+final class ForgotPasswordAction
 {
     public function __invoke(ForgotPasswordRequest $request, string $flavor = 'api'): JsonResponse|RedirectResponse
     {