+++ src/Http/Actions/ListProvidersAction.php
@@ -2,12 +2,12 @@
 
 declare(strict_types=1);
 
-namespace Technobase\AuthKit\Http\Callbacks;
+namespace Technobase\AuthKit\Http\Actions;
 
 use Illuminate\Http\JsonResponse;
 use Technobase\AuthKit\Facades\AuthKit;
 
-final class ProvidersCallback
+final class ListProvidersAction
 {
     public function __invoke(): JsonResponse
     {